moved all drag drop code parts to the DragDropProxy.cs class
